requirements.txt
streamlit>=1.20.0
pymongo[srv]>=4.0.0
python-dotenv>=1.0.0
pandas>=1.5.0
openpyxl>=3.1.0
dnspython>=2.3.0
